generalized code for running in the cloud

+++ train.py
@@ -18,7 +18,7 @@ from sklearn.model_selection import train_test_split
 
 from predict import predict
 from load_preprocess_data import load_data, preprocess
-from utility import make_dir, clean_path
+from utility import make_dir
 from plot import plot_corrected_depths
 
 
@@ -68,14 +68,14 @@ def _current_time():
 def load_preprocess_data(bed_file, fasta_file, chromosome_number, region_start, region_end):
     data = load_data(bed_file, fasta_file, chromosome_number, region_start, region_end)
 
-    data_train, data_test = train_test_split(data, test_size=0.2, random_state=42)
+    data_train, data_dev = train_test_split(data, test_size=0.2, random_state=42)
 
     # noinspection PyTypeChecker
-    return (data_train,) + preprocess(data_train) + (data_test,) + preprocess(data_test)
+    return (data_train,) + preprocess(data_train) + (data_dev,) + preprocess(data_dev)
 
 
 def _writer(tensorboard_dir, sub_dir_name, session):
-    return tf.summary.FileWriter(clean_path(tensorboard_dir) + '/' + sub_dir_name, session.graph)
+    return tf.summary.FileWriter(os.path.join(tensorboard_dir, sub_dir_name), session.graph)
 
 
 def mini_batches(images, observed_depths, graph, batch_size=256):
@@ -96,7 +96,7 @@ def mini_batches(images, observed_depths, graph, batch_size=256):
                graph.y: shuffled_observed_depths[batch_indices, :]}
         batch_indices += batch_size
 
-    # Omit the last mini-batch for simplicity.
+    # Omitting the last mini-batch for simplicity.
     # Including it MAY require scaling the learning rate by the size of the mini batch
     # min_index = min(batch_indices)
     # yield {graph.X: images[min_index:, :, :, :],
@@ -104,8 +104,7 @@ def mini_batches(images, observed_depths, graph, batch_size=256):
 
 
 def train(number_epochs, logging_interval, checkpoint_interval,
-          print_to_console, tensorboard_dir, graph_variables_dir, training_data_dir):
-
+          print_to_console, train_data_dir, trained_model_dir):
     # !!! use tf.data API when input data is distributed across multiple machines !!!
     # !!! https://www.youtube.com/watch?v=uIcqeP7MFH0 !!!
 
@@ -115,12 +114,12 @@ def train(number_epochs, logging_interval, checkpoint_interval,
         sys.exit()
 
     (data_train, images_train, observed_depths_train,
-     data_test, images_test, observed_depths_test) = load_preprocess_data(
-        bed_file='../data/facnn-example.regions.bed.gz',
-        fasta_file='../data/human_g1k_v37.fasta',
+     data_dev, images_dev, observed_depths_dev) = load_preprocess_data(
+        bed_file=os.path.join(train_data_dir, 'facnn-example.regions.bed.gz'),
+        fasta_file=os.path.join(train_data_dir, 'human_g1k_v37.fasta'),
         chromosome_number='1',
         region_start=0,
-        region_end=2000000)
+        region_end=20000)
 
     _, image_width, image_height, _ = images_train.shape
     graph = Graph(image_width, image_height)
@@ -130,13 +129,16 @@ def train(number_epochs, logging_interval, checkpoint_interval,
     with tf.Session() as session:
         session.run(tf.global_variables_initializer())
 
-        import shutil
-        shutil.rmtree(tensorboard_dir)
+        tensorboard_dir = os.path.join(trained_model_dir, 'tensorboard')
+
+        if os.path.isdir(tensorboard_dir):
+            import shutil
+            shutil.rmtree(tensorboard_dir)
 
         # using a context manager, i.e. the "with" syntax,
         # ensures that the FileWriter buffer is emptied, i.e. data are written to disk
         with _writer(tensorboard_dir, 'train', session) as train_writer, \
-                _writer(tensorboard_dir, 'test', session) as test_writer:
+                _writer(tensorboard_dir, 'dev', session) as dev_writer:
 
             start_time = _current_time()
             for epoch in range(number_epochs):
@@ -148,9 +150,9 @@ def train(number_epochs, logging_interval, checkpoint_interval,
                 cost_train, predicted_depths_train = session.run(
                     [graph.cost, graph.predictions], feed_dict=feed_dict_train)
 
-                feed_dict_test = {graph.X: images_test, graph.y: observed_depths_test}
-                cost_test, predicted_depths_test = session.run(
-                    [graph.cost, graph.predictions], feed_dict=feed_dict_test)
+                feed_dict_dev = {graph.X: images_dev, graph.y: observed_depths_dev}
+                cost_dev, predicted_depths_dev = session.run(
+                    [graph.cost, graph.predictions], feed_dict=feed_dict_dev)
 
                 if epoch % logging_interval == 0:
                     if print_to_console:
@@ -158,72 +160,78 @@ def train(number_epochs, logging_interval, checkpoint_interval,
                         print('epoch:', epoch,
                               'elapsed time (secs):', _current_time() - start_time,
                               'cost_train:', cost_train,
-                              'cost_test:', cost_test,
+                              'cost_dev:', cost_dev,
                               'observed_depths_train:', observed_depths_train[:number_depths, 0],
                               'predicted_depths_train:', predicted_depths_train[:number_depths, 0])
 
                     # https://stackoverflow.com/questions/31674557/how-to-append-rows-in-a-pandas-dataframe-in-a-for-loop
                     # https://stackoverflow.com/questions/37009287/using-pandas-append-within-for-loop/37009377
-                    log.append({'epoch': epoch, 'cost_train': cost_train, 'cost_test': cost_test})
+                    log.append({'epoch': epoch, 'cost_train': cost_train, 'cost_dev': cost_dev})
 
                     # https://www.tensorflow.org/guide/summaries_and_tensorboard
                     train_writer.add_summary(session.run(graph.all_summaries, feed_dict=feed_dict_train), epoch)
-                    test_writer.add_summary(session.run(graph.cost_summary, feed_dict=feed_dict_test), epoch)
+                    dev_writer.add_summary(session.run(graph.cost_summary, feed_dict=feed_dict_dev), epoch)
 
                 if epoch % checkpoint_interval == 0:
-                    graph_variables_checkpoint = {
+                    graph_variables = 'graph_variables'
+                    graph_variables__trained_model_checkpoint = graph_variables + '/trained_model.ckpt'
+                    tf.train.Saver().save(session, os.path.join(trained_model_dir,
+                                                                graph_variables__trained_model_checkpoint))
+                    checkpoint = {
                         # https://stackabuse.com/tensorflow-save-and-restore-models/
                         # https://cv-tricks.com/tensorflow-tutorial/save-restore-tensorflow-models-quick-complete-tutorial/
                         # the second argument of save() is the filename prefix of checkpoint files it creates
-                        'prefix': tf.train.Saver().save(session,
-                                                        clean_path(graph_variables_dir) + '/trained_model.ckpt'),
-                        'graph_variables_dir': graph_variables_dir}
+                        'prefix': graph_variables__trained_model_checkpoint,
+                        'graph_variables_dir': graph_variables}
                     import json
-                    with open('checkpoint.json', 'w') as fp:
-                        json.dump(graph_variables_checkpoint, fp, indent=4)
+                    with open(os.path.join(trained_model_dir, 'checkpoint.json'), 'w') as fp:
+                        json.dump(checkpoint, fp, indent=4)
 
                     data_train['predicted_depth'] = predicted_depths_train
-                    data_test['predicted_depth'] = predicted_depths_test
+                    data_dev['predicted_depth'] = predicted_depths_dev
                     pickle(data_train.sort_values('start'),
-                           data_test.sort_values('start'),
-                           pd.DataFrame(log),
-                           training_data_dir)
+                           data_dev.sort_values('start'),
+                           pd.DataFrame(log), trained_model_dir)
 
 
 def _generate_file_names(dataframe_dir):
     make_dir(dataframe_dir)
-    return (clean_path(dataframe_dir) + '/data_train.pkl',
-            clean_path(dataframe_dir) + '/data_test.pkl',
-            clean_path(dataframe_dir) + '/log.pkl')
+    return (os.path.join(dataframe_dir, 'train.pkl'),
+            os.path.join(dataframe_dir, 'dev.pkl'),
+            os.path.join(dataframe_dir, 'log.pkl'))
 
 
-def pickle(data_train, data_test, log, dataframe_dir):
-    data_train_filename, data_test_filename, log_filename = _generate_file_names(dataframe_dir)
+def pickle(data_train, data_dev, log, dataframe_dir):
+    data_train_filename, data_dev_filename, log_filename = _generate_file_names(dataframe_dir)
     data_train.to_pickle(data_train_filename)
-    data_test.to_pickle(data_test_filename)
+    data_dev.to_pickle(data_dev_filename)
     log.to_pickle(log_filename)
 
 
 def unpickle(dataframe_dir):
-    data_train_filename, data_test_filename, log_filename = _generate_file_names(dataframe_dir)
+    data_train_filename, data_dev_filename, log_filename = _generate_file_names(dataframe_dir)
     return (pd.read_pickle(data_train_filename),
-            pd.read_pickle(data_test_filename),
+            pd.read_pickle(data_dev_filename),
             pd.read_pickle(log_filename))
 
 
 def main():
-    training_data_dir = '../trained_model/training_data'
-    train(number_epochs=100, logging_interval=1, checkpoint_interval=1, print_to_console=True,
-          tensorboard_dir='../trained_model/tensorboard', graph_variables_dir='../trained_model/graph_variables',
-          training_data_dir=training_data_dir)
+    import argparse
+    parser = argparse.ArgumentParser()
+    parser.add_argument('--train_data')
+    parser.add_argument('--trained_model')
+    args = parser.parse_args()
 
-    training_data, validation_data, training_log = unpickle(training_data_dir)
+    train(number_epochs=10, logging_interval=1, checkpoint_interval=1, print_to_console=True,
+          train_data_dir=args.train_data, trained_model_dir=args.trained_model)
+
+    train_data, dev_data, train_log = unpickle(args.trained_model)
 
     from plot import compute_observed_depth_mean
-    observed_depth_mean = compute_observed_depth_mean(training_data)
+    observed_depth_mean = compute_observed_depth_mean(train_data)
 
-    plot_corrected_depths(training_data, observed_depth_mean, title='training data')
-    plot_corrected_depths(validation_data, observed_depth_mean, title='validation data')
+    plot_corrected_depths(train_data, observed_depth_mean, title='train data')
+    plot_corrected_depths(dev_data, observed_depth_mean, title='dev data')
 
 
 if __name__ == '__main__':
